fix(codegen): assert guard tests only over verified fixtures (#312) (#315)
Generated guard tests asserted a truth value over a fixture the solver
had searched for but never proved. The falsifier's Cmp case handles
only 'field <op> literal', so a field-vs-field guard
('member_index < member_count') fell through to string heuristics and
could return an assignment leaving the guard TRUE. Asserting
'!guard(...)' over it produced a generated test that fails against
correct code — six such failures across the bundled examples.

Add a verification boundary rather than redesigning the search: build
the exact fixture the test will contain, evaluate the guard against
it, and emit a truth-value assertion only when evaluation confirms it.
The outcome is explicit:

- Solved       — evaluation confirmed the wanted value; assert.
- Contradicted — evaluation produced the opposite; emit the fixture
                 with an explanation instead of an assertion.
- Unsupported  — the guard reads accounts, containers, or operators
                 outside the decidable fragment; the fixture is
                 unverified either way.

Deliberately not #294's 'Unsatisfiable': proving no witness exists
needs a real solver. What this boundary decides is whether the witness
in hand actually works. A better search (#263) still benefits — it
only changes how often the outcome is Solved.

The evaluator covers literals, state fields, params, consts,
comparisons, boolean operators, and checked arithmetic; everything
else is Unsupported by construction.

Also fixes the same u8-subscript defect as #295 (c) in the Kani
conformance emitter: 's.voted[member_index] == 1' interpolated the raw
effect target, an E0277 that only surfaced once the gate's Kani
compile step began running for multisig.

Result: lending 22/22 and multisig 51/51 generated unit tests pass,
and the generated-artifact gate is fully green for the first time.
13 real guard assertions are kept across the two examples; each
suppression carries its specific reason.

Regression: guard_assertions_only_over_verified_fixtures asserts every
guard test either asserts or explains its suppression, never neither.

+++ examples/rust/multisig/programs/tests/kani.rs
@@ -1067,7 +1067,7 @@ fn verify_approve_effect_voted_member_index() {
     let pre_rejection_count = s.rejection_count;
     if approve(&mut s, member_index) {
         assert!(
-            s.voted[member_index] == 1,
+            s.voted[member_index as usize] == 1,
             "voted[member_index] must equal 1"
         );
         assert!(
@@ -1163,7 +1163,7 @@ fn verify_reject_effect_voted_member_index() {
     let pre_rejection_count = s.rejection_count;
     if reject(&mut s, member_index) {
         assert!(
-            s.voted[member_index] == 1,
+            s.voted[member_index as usize] == 1,
             "voted[member_index] must equal 1"
         );
         assert!(
@@ -1365,7 +1365,7 @@ fn verify_add_member_effect_members_member_index() {
     let pre_rejection_count = s.rejection_count;
     if add_member(&mut s, member_index, member_pubkey) {
         assert!(
-            pubkey_eq(&s.members[member_index], &member_pubkey),
+            pubkey_eq(&s.members[member_index as usize], &member_pubkey),
             "members[member_index] must equal member_pubkey"
         );
         assert!(
